implemented ble gatt server. disabled hrfco
UV data is sent every 0.5 seconds. the led is toggled on then off in the
same intervals.
hrfco has been disabled as ble wouldn't work with it enabled. the
default behavior of having hrfco enabled via the project configurator,
causes the radio peripheral to be clocked by the hrfco instead of hfxo.
since the radio peripheral is expecting 38.4 mHz clock from hfxo, but is
given a different value from hrfco, radio would not work.