removed letimer/gpioint. bt stack soft timer used. uv data buffer.
letimer is replaced with the bt stack soft timer because the soft timer
is already integrated with the bt stack. gpioint no longer used, gpio
interrupts have been implemented without that component. uv data is now
buffered with a size of 3. on the 3rd data measurement, ble notification
is sent with the data of the entire buffer.